CHANGE: renamed `/rest` => `/api`
