+++ README.md
@@ -59,7 +59,7 @@ Edit `.barreslby.json` to customize it:
 
 ```json
 {
-  "directory": ["./src/controllers/rest", "./src/controllers/pages"],
+  "directory": ["./src/controllers/api", "./src/controllers/pages"],
   "exclude": ["__mock__", "__mocks__", ".spec.ts"],
   "delete": true
 }
